pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anything-extractor"
version = "0.1.0"
requires-python = ">=3.11"
description = "Self-adaptive data extraction system for professional PDF documents"
dependencies = [
    "typer>=0.9",
    "openai>=1.0",
    "sqlalchemy>=2.0",
    "alembic>=1.12",
    "psycopg2-binary",
    "redis>=5.0",
    "pymupdf>=1.23",
    "pydantic>=2.0",
    "pydantic-settings",
    "openpyxl",
    "rich",
    "gitpython",
    "httpx",
]

[project.scripts]
ae = "ae.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
